mudanças em alguns ficheiros

+++ include/tag.h
@@ -9,6 +9,6 @@ TAG create_tag();
 gchar* get_key_tag_name(TAG t);
 long get_id_tag(TAG t);
 void set_key_tag_name(TAG t, char* str);
-void set_id_tag(TAG t, int id_tag);
+void set_id_tag(TAG t, long id_tag);
 void free_tag(TAG t);
 #endif
